refactor(webapp): move billing store to runes
